CL language specifics as TS
